refactor: Optimized existing tests
