+++ tests/Integration/Controller/Admin/Area/DispatchArea/DispatchAreaShowControllerTest.php
@@ -17,15 +17,16 @@ class DispatchAreaShowControllerTest extends AppWebTestCase
 
     public function testAdminsCanViewDispatchAreas(): void
     {
-        UserFactory::new(['username' => 'admin'])->asAdmin()->create();
-
+        // Arrange
         StateFactory::createOne();
         SupplyAreaFactory::createOne();
+        DispatchAreaFactory::createOne();
 
-        $dispatchArea = DispatchAreaFactory::createOne();
+        $dispatchArea = DispatchAreaFactory::random();
 
+        // Act& Assert
         $this->browser()
-            ->loginAs('admin', 'password')
+            ->actingAs(UserFactory::new()->asAdmin()->create()->object())
             ->visit('/admin/area/dispatch/'.$dispatchArea->getId())
             ->assertSuccessful()
             ->assertSeeIn('h2', 'Dispatch Area')
