+++ tests/Integration/Controller/Admin/Area/State/StateEditControllerTest.php
@@ -15,12 +15,12 @@ class StateEditControllerTest extends AppWebTestCase
 
     public function testAdminsCanEditAState(): void
     {
-        UserFactory::new(['username' => 'admin'])->asAdmin()->create();
-
+        // Arrange
         $state = StateFactory::createOne();
 
+        // Act& Assert
         $this->browser()
-            ->loginAs('admin', 'password')
+            ->actingAs(UserFactory::new()->asAdmin()->create()->object())
             ->visit('/admin/area/state/'.$state->getId().'/edit')
             ->assertSuccessful()
             ->assertSeeIn('title', 'Edit State')
