+++ tests/Integration/Controller/Security/SecurityControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration\Controller;
+namespace App\Tests\Integration\Controller\Security;
 
 use App\Factory\UserFactory;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
@@ -18,8 +18,10 @@ class SecurityControllerTest extends WebTestCase
 
     public function testYouCanLoginAndLogout(): void
     {
+        // Arrange
         UserFactory::new(['username' => 'foo'])->create();
 
+        // Act& Assert
         $this->browser()
             ->visit('/login')
             ->assertSeeIn('title', 'Login')
